requirements.txt
absl-py==2.1.0
boto3==1.34.154
botocore==1.34.154
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.13.1
filetype==1.2.0
fire==0.6.0
fonttools==4.53.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pybboxes==0.1.6
pyclick==0.0.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.37
s3transfer==0.10.2
sahi==0.11.18
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==70.0.0
shapely==2.0.5
six==1.16.0
smmap==5.0.1
stable_baselines3==2.3.2
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.73
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
xlib==0.21
yolov5==7.0.13
